requirements.txt
Flask==2.0.1
Pillow==9.0.0
requests==2.26.0
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
opencv-python==4.5.3.56
numpy==1.21.2 